AudioPluginService: wait for debugger only if it is enabled.
Service is debuggable only if `android.os.Debug.waitForDebugger()` is
invoked, and it just stops there forever if the process is not attached by
debugger. It is problematic if we are trying to debug service that can be
used either as a debugged service or not-debugged service.

Therefore this new optional field is introduced to make it explicitly
debuggable. The `MainActivty` in `aappluginsample` shows how it is used.