Now samples are simplified and more consistent to become comparable.
We have to find how remote plugins don't work, and to do so it's easier
to have both local plugin sample (which works well) and remote sample
(which doesn't) to do the same job and differ in the processing model.

Therefore, so far both samples deal with only one plugin to be loaded
and processed, and both samples have L/R channels to be processed
appropriately.