LV2 dependency sources are now duplicated before building for each ABI.
The build script was kind of lame because only the first ABI gets clean
build. cerbero does handle this cleanly so we follow that.

It would have been cleaner if we can premise git repositories (then we
can use worktree), but that would force git existence and blocks build
from source archives.