Adding service connection feature in AudioPluginHost and natives.
Connections are in practice binders. They are stored in native aap::binders.
Plugins are getting isolated from services and will be managed differently.
The instantiation process will have to be adjusted to appropriately use
the shared binder across multiple plugin instances.

At current state, plugin is instantiated and saple runClientAAP() is invoked
but there is no adjustment on passing instance and it therefore causes some
SEGV.