Fix Android plugin query for JUCE apps.
There was a simple problem that the JUCE-AAP way of querying plugin list
actually didn't work, due to the fact that the latest querying model is
based on bindService() results. Only bound services were listed as the
plugin containers.

It was obviously not working, but it had been hidden long time due to
the fact that JuceAAPAudioPluginHost app kept plugin list cache and
that had never gone away, resulting in that eventually consistent list
of plugins on the debuggee. Of course this does not work on any new
setup, meaning that it only worked on my development emulator.