A bunch of renaming.
Here are some naming decisions made and applied in these changes.

(1)PluginInformation

It was "descriptor" before, and now it is information, while there were
still a lot of pieces of code that has "descriptor" in members and fields.
There should be solid unification on which is being adopted:

- PluginInformation
- PluginDescriptor

The latter sounded more like a structure with some handles (pointers) like
`FileDescriptor` and `lv2_descriptor`.

PluginInformation collides juce::PluginInformation but the name is chosen
based on the comparison above.

(2)PluginServiceConnection

It was renamed from AudioPluginServiceConnection. When renaming, there were
3 candidates:

- AudioPluginServiceConnection
- PluginServiceConnection
- ServiceConnection

The class is actually derived from `android.content.ServiceConnection` so
the last choice should be avoided unless there are any other strong reasons
to dare to. AudioPluginServiceConnection was a bit too verbose, and it
gives somewhat public feeling, and since there is no other plugin systems
involved, there won't be any confusion to just name it as `plugin`
service conn.

A remaining FIXME is left that this class should be actually non-public.

(3) serviceIdentifier -> pluginPackageName and pluginLocalName

There used to be "serviceIdentifier" which indicates an unique ID string
that indicates both Android service name and a class name to indicate
a particular Service class. Since we need both package name and class name
almost everywhere, it is better to just require those two keys at a time.