aaphostsample: Fixed inconsistent parameter list passing.
The parameter value list should be initialized appropriately at constructor
as well as set value appropriately at getView(), as getView() is invoked
only when the view instance is created. And whenever it went out of scope,
the value was reset to default, which is unexpected.

(The changes contain disabled code for Kotlin implementation of audio
processing in AAPSampleRemoteInterop_native.cpp, but it still does not
work yet.)