Remove EditorInstance (and reorder some code).
As part of UI integration, I noticed that we should not assume that a
PluginInstance is in the same code domain as that of EditorInstance.
Like LV2, it is possible that we have separate native library loader
that does not really share strongly typed API between plugin and UI.
Some generic UI loaders that only takes parameters in manifest may show
users the corresponding UI controls.

We are still unclear on how plugin and editor communicates each other,
but so far it makes more sense to remove the existing EditorInstance API.