fix build; native dependency resolution in AGP 7.0 era.
With the latest AGP, adding build dependency on mergeDebugNativeLibs does
not seem to work anymore. They have been making breaking changes in build
tasks, and now the relevant build task for CMake seems to be buildCMakeDebug.

Odd is that there is no buildCMakeRelease, and therefore we cannot depend
on such non-existent build task at all(!). I have no idea if it is a bug
or not. (how come it is not though?)