Revert "midi-device-service: remove munmap() call."
This reverts commit b36ec5af428c871fe1274b6e56576730f90e1bea.

It was not that consistent; I'm still getting the close() fdsan failure.
It is hard to find out consistent behavior because Android Studio is quite
buggy so that it does not reflect the latest native code changes at debugging
(clearing storage makes it consistent, but it should not be required
every time we debug app).

The only use case of ASharedMemory_create() is this ndk-samples app
https://github.com/android/ndk-samples/blob/87ae9bff54a9264e8b92515bfbda055d4ebe28bc/nn-samples/basic/src/main/cpp/simple_model.cpp
and it calls munmap(), so it was presumably correct to invoke it.