make AAPMidiProcessor midi processing consistent between midi1 and midi2.
Also removed unused code for clarity.

update cmidi2.h for consistent Set New Protocol detection.

AAPInstrumentSample also detect Set New Protocol, but it is (so far) not
used when it is connected as a MidiDeviceService (because it is always
MIDI2 at instantiation time).

It's still not working; it seems to drop note on, maybe some sort of
atomic locking is required i.e. prevent reading when it's writing and
vice versa.