bump AGP to D-alpha02, compose (1.1.0) and kotlin (1.6.10).
They are actually chained update requirements...

AGP is always required to be the latest, as long as we use Android Studio
from Canary (not if we use AS from beta).
AGP Dolphin-alpha02 this time somehow bumped the minimum Compose dependencies
to 1.1.0 or whatever around, regardless of Compose Gradle Plugin (I didn't
touch its version this time. I have no idea how this requirement bump is
even possible, implementation wise).
And by bumping Compose versions to 1.1.0, Kotlin version requirement is
bumped to 1.6.10 (which is actually what I wanted for long time, I was
bound to 1.5.31 forever).