Another big host API revamp: client asyncification.
context: https://github.com/atsushieno/android-audio-plugin-framework/issues/102

Our instance creation at client side needs to be non-blocking because AAP
is based on Service and Binder, and `ServiceConnection.onServiceConnected()`
is not invoked if the main thread is blocked. Current native host API was
designed to be blocking, thus it needs to be fixed.

This revamps the overall `PluginClient` implementation to work in
asynchronized manner, which requires whole redesign.

As commented on issue #102, AUv3 is based on asynchronous design too, which
makes sense. The corresponding aap-juce implementation is updated too to
at least compile AudioPluginHost. Though it crashes when it comes back to
callback, so we need further fixes.