extension service: Intern URI strings like LV2 URID, for AAPXS instance wrapper.
We cannot simply lookup values in map<const char*, V> by [] operator.
They have to be first mapped to stable int value for immediate comparison.
We can still call `strcmp()` in every lookup, but that costs a lot.
That's what LV2 URID is for. Or "Intern" string in .NET System.String or
System.Xml.NameTable.

While it's not really in use (yet), we offer interned lookup only mode
so that only interned strings can be used for looking up entries.
It is used only when `onlyInterned` argument is `true`.