extension service: fixed various memory allocation issues.
Now we reached to the stage where extension() call is made. However we fail
to allocate shared memory (as we pass 0 as size when binder-client-as-plugin
calls ASharedMemory_create(), it ignores the allocation part).

Actually the AAPXS framework currently lacks how to initialize and allocate
the extension data structure at host side. We need to come up with some ideas.