aapxs: replace the "host extension" concept with "extension data".
In earlier plans, I kept saying like "host extensions will be removed as
the concept is wrong". It is true, but we cannot simply remove it because
we still need something exactly like that for plugin extensions.

For example, the MIDI2 extension provides configuration for the MIDI
message contents whether they are MIDI1 or MIDI2 (if we keep things so).
The information is not 100% plugin matter; the host might be able to deal
only with 1.0, or even only with 2.0(!) Therefore, what host sends to
client in the initial extension data would matter.

It is actually interpreted by aapinstrumentsample to determine how to
deal with the MIDI messages there. That's what (ex-)MIDI-CI extension did.

Thus, `AndroidPluginHost` argument in `instantiate()` remain, but the
`get_host_extension()` is renamed to `get_extension_data()` to retrieve
the shared data pointers as in the extension type data, to make it possible.