Temporarily change minSdkVersion to 30 to use newer NDK r24.
As it was sometimes mentioned in this repo, NDK and/or SDK tools has an
issue where NdkBinder header files are broken for minSdk 29 (the headers
have unconditional reference to API Level 30).

I had been avoiding the issue to stick to the old NDK r21, but the
debugging experience on Android Studio has been inconsistent and annoying,
and I cannot be confident as long as I use the outdated version of NDK.

We cannot really lower the minSdkVersion to 29 for production, that is
definitely not our plan. But it would be earlier that Google fixes the
tooling issue than we reach the point where minSdkVersion seriously matters.
So, **temporarily** change the minSdkVersion to 30.

Fix build that was not failing with NDK r21 as well.