Explicitly exclude libc++shared.so from packaging (workaround NDK bug).
context: https://github.com/google/prefab/issues/127

I would call it an NDK bug that AGP packages libc++_shared.so, which
only leads to ODR violation. It should be app build that resolve
libc++_shared.so and packaging libc++_shared.so in an aar is stupid.
It's like bundling mscorlib.dll in *.nupkg, or bundling rt.jar in our
*library*. It should be *.aab or *.apk that should package it.

While it is AGP which we app developers face this issue, it is not AGP
which is faulty; they are not at the position to determine the design on
how native packaging works.

In any case, all we could do for now is to exclude libc++_shared.so from
packaging.