Switch to prefab.
This resolves #57 once it is merged to main.

It was meant to resolve #109 to see if prefab packaging fixes the issue,
but it didn't work. The latest prefab aars contains jniLibs just like
non-prefab aars, along with `modules` subdirectory.

It didn't work, but since the prefab packages do not seem to block
interaction with non-prefab packaging strategy, we can now switch to
prefab packages.