Add AAPXSClientInstanceManager class.
It is to isolate AAPXS client instance management job from RemotePluginInstance.

RemotePluginInstance is a native hosting feature, and hence core of
libandroidaudioplugin hosting implementation.
For other implementations (e.g. we also have Kotlin hosting API), we still
want to manage AAPXS, and then we need something that is independent of the
native hosting implementation.

Though it is not achieved yet, because AAPXSClientInstanceWrapper depends on
RemotePluginInstance. We need further isolation.