initial work on new client API for Kotlin.
The new client is based on aap::RemotePluginInstance, not direct use of
AIDL binder proxy. For details, see issue #112.

There is now aap::PluginClient::createInstance() which is the sync version
which eases instancing. Kotlin host already connected to the plugin when
it is instantiating a plugin, so async work is extraneous.