rename PortInformation `index` to `id` and access port list by index, not id.
It is going to be required to support dynamically populated ports as per
`NEW_PORTS.md` design doc. There are some lines commented out with FIXME
that actually instantiates so called "system common" ports.

We need further changes to support configurePorts() equivalents in AAP
service, and that is going to require changes to AIDL, which should come
next or around...