bump Intent Action (breaking) for split prepare step.
The actual necessary changes for dynamic ports are still not achieved yet.

We first need to set up configured ports for both client and service.
For system common ports, they are done automatically in common code now
(`startPortConfiguration()`). So far we all `setupPortsViaMetadata()` to
actually set up ports (as we haven't done with the actual ports config),
but that will be conditionally applied once it's all done.

To achieve that, we kinda have to split prepare() AIDL transaction to
beginPrepare() and endPrepare() (we could hack around to stick to old API,
but that would bring in unnecessary complication, so we chose simplification).

Since it is a breaking change, the intent action is now bumped to "V2"
and older AAPs will remain as older ecosystem (if I'd say so).