Revert "rename PortInformation `index` to `id` and access port list by index, not id."
This reverts commit c25defc03caf8518a802d8f37418705771cd81c4.

context: https://github.com/atsushieno/android-audio-plugin-framework/issues/44#issuecomment-1211512985

The primary reason for the revert is that we need way more complicated
changes that would also complicate public plugin API - the plugin API
is based on array based audio/MIDI buffers just by port index, not by ID.
Dynamic mapping of the ports needs various rewrites, and even if we did that
we are not very clear on what we could bring in as the benefits.

The basic defective idea is mostly about negative ports that prevent
index-based access to buffers. They could be either appended to the last
port, or simply unbandled from user port buffers.