Some .V2 breaking changes in MIDI port for the next PluginPreview MIDI support.
We are still not-working state for default port configuration for
instruments because it only populates MIDI2 port, not MIDI port, and
our PluginPreview only emits MIDI1 events for the preview template
sequence. It should conditionally emit MIDI2 sequence.

Therefore, aapinstrumentsample kind of reverts the latest code and
manually configured ports are back for now. They will be removed once
we are done with the work.

In the meantime, the current code base is quite annoying to support
both MIDI 1.0 and MIDI 2.0, especially if we support arbitrary sequence.
Thus added MML support by mugene-ng (that I manage development) and
ktmidi to programatically generate those sequences from one single MML.

We still need to cover absolute time conversion for splitting sequence
per process block, and what mugene-ng generates is not suitable (as it
contains tempo state). We will have to deal with it next.