aapinstrumentsample provides only MIDI2 ports now.
It will be consistent with the implicit ports that will be populated by AAP
when there is no declaration in the plugin.

At this state, both parameter changes and other MIDI events such as note on
will be supported only via MIDI2 ports. Like CLAP project states, multiple
event inputs (ports) are considered harmful (because of potentially
inconsistent order of processing of them), and supporting *both* MIDI1 and
MIDI2 in one single plugin is not realistic.

samples-host-engine implements minimal testing support for MIDI2 ports,
so as midi-device-service by translating MIDI1 inputs to UMPs.