Parameter changes in UMP is now sysex8, not Assignable Controller.
context: https://github.com/atsushieno/android-audio-plugin-framework/issues/116#issuecomment-1307491933

If we use Assignable Controllers (NRPN) to manipulate parameters, we will run
into an issue: the plugin will not be able to receive NRPNs anymore.

Assignable Controller message in UMP is new, but MIDI1 NRPN messages via
Control Changes (NRPN and DTE) are supposed to be translated to them, which
means that MIDI1 inputs to AAP (which is to be translated to UMPs at the
entry gate) will be thrown into conflict: what is this AC? For parameters?
or translated NRPN message? The latter is most likely just killed.

This problem is just like how VST3 killed control changes: https://forums.steinberg.net/t/vst3-and-midi-cc-pitfall/201879 (though the impact is smaller)

This will kill synths like traditional MIDI device emulator (like Roland
Sound Canvas VA).

Actually I once prepared for the possible conflicts and had an unused
property (or function or whatsoever) in Parameters extension. There was
`aap_parameters_mapping_policy` which indicates that the mapping is
straightforward AC, or AC with the flipped top bit, implying that no one
would use parameter index >32768. But it's still not a safe guess.
Any plugin could use parameter index identity like that too.

Therefore, this changeset contains an anternative fix: use sysex8 instead.

The actual sysex8 format is described in `parameters.h`. And since syse8
does not contain "channel" and "key" (note), they are added as part of
the sysex8 message content. It is still put within a 16 bits packet.