Switch to coroutines model for plugin service connection stuff.
We should be cleaning up ugly parts of the hosting API.
The async connection stuff is complicated mostly due to `bindService()` and
I left it classic way, but they can be better.
A remaining annoyance is JNI interop part (as it's not realistic to take
suspend functions into consideration in C++ callables) so that part is
not asynchronous.