Avoid creating AIDL proxy when aiBinder instance is the same as former ones.
This fixes issue #123.

Without this fix, AudioPluginNatives.addBinderForClient() was causing:

> Class cannot be associated on object which already has a class.

... which is in AIBinder::associateClass(), at https://cs.android.com/android/platform/superproject/+/master:frameworks/native/libs/binder/ndk/ibinder.cpp;l=126;drc=b0558503af653c3872b91fed3c685cae092876fc;bpv=1;bpt=1

It is called by `aidl::org::androidaudioplugin::IAudioPluginInterface::fromBinder()`.

It happens when the context `AIBinder` is already bound to another AIDL
interface class. I'm not sure if that is a valid behavior, but to match
the expected behavior, we store AIBinder and AIDL proxy and reuse them if
another request to the same AIBinder occurs.

(I tried to use jobject binder as the cache, but it seems that the jobject
is actually IBinderProxy and the instance is identical for different
AIDL interface classes. They get differentiated only at AIBinder.)

With this fix, plugin instancing tests finally pass without limiting the
test cycle number as well as the number of parallel instances.