parameters MIDI mapping policy can be retrieved by client too.
In other words, it wasn't before.

It is due to different behaviors between (local) service and (remote) client.

Basically MIDI mapping policy is stored per plugin service, and it does not
make sense if the policy value is stored at a host (hosts are free to send
whatever, regardless of the shared preference value, but it does not affect
how the plugin actually behaves).

Thus, for client, the flag must be retrieved via binder. Though it's not
going into the plugin extension; it is androidaudioplugin.aar (including
libandroidaudioplugin.so) that queries the settings and return to the client.

While implementing this, another annoying issue emerged; we could not resolve
`AudioPluginMidiSettings` class from AudioPluginService that was instantiated
remotely - the identical code that worked for AudioPluginService that was
instantiated from within the same app (PluginPreview within the same app)
did not work for remotely-instantiated ones.

It was due to insufficient `SystemClassLoader` that was unable to resolve
classes in the apk. It has to be resolved via `Context.classLoader`.

To achieve that, there are additional lines to retrieve
`AudioPluginMidiSettings` class via reflection trick within JNI.

They are all annoying changes, but had to be made.
The code is not efficient either, but hopefully does not matter.
We don't call it often.