AAPXS now has "is it mandatory?" flag. Implement MIDI mapping policy with it.
Apparently the previous change that claimed that "it had implemented MIDI
mapping for clients" was not ture... the actual implementation is done here.
aaphostsample now successfully reflects what the plugin actually specifies
for the MIDI mapping policy.

This change contains the solution for https://github.com/atsushieno/android-audio-plugin-framework/issues/132 -
the parameters extension is usually not implemented by the plugins i.e.
`get_extension()` does not return for it. But host should get the mappings
that are merely stored as shared preferences at AudioPluginService side.

Thus, AAPXS now has `is_implementation_mandatory` field that indicates
whether the extension works without `get_extension()` results.
The extension still has to be declared at `aap_metadata.xml` (otherwise
AAPXS will not be instantiated at all).

With this new feature, parameters extension is implemented as "optional".

In practice, the service implementation only invoked the JNI function
that retrieves the shared preference value, without invoking the service's
`get_mapping_policy()` (so far).

In parameters extension, the previous `get_mapping_policy()` in the host
extension is renamed to `get_user_mapping_policy()` as it is now confusing
with that of the plugin extension.