finally implemented MIDI mapping feature in AudioPluginMidiDeviceService.
Now we can change parameters via CC, ACC and Per-Note ACC by translating
them to sysex8, depending on the MIDI mapping settings.
Similarly, program changes are consumed (if mapped) as preset changes.

Though I'm reconsidering the overall design; it would make a lot more sense
if the plugins declare what they expect to be passed directly, and
AAP MidiDeviceService does not translate them to parameter changes and
program changes so that they can consume as is.