Changed the MIDI mapping policy semantics.
It is now a set of flags for a plugin to declare which MIDI messages are
directly consumed by the plugin and therefore cannot be used for other
purposes (namely, MIDI mappings).

AAP MidiDeviceService will not translate those messages that are indicated
for direct consumption by plugin to parameter changes or preset changes.

The actual behavior is not very different, but the semantics of the flags
as well as the default value have changed.

With this new design, we don't really need "user mapping policy", so
`get_user_mapping_policy()` is removed from parameters host extension.
We keep things simple as possible as we can.