Finally in-plugin-process UI shows up on aaphostsample.
It is based on a new ALooper basis. Since the binder extension message for
GUI extension should not run GUI stuff by itself, it should first dispatch
the request to another non-realtime message loop. It is done with ALooper.

Since we want the thread correctly handle Java classes, we first create
a java.lang.Thread, and run its loop in native code.

We should probably use this foundation for any non-realtime request.