isolate aap::PluginHost from aap::PluginInstance (for each client/service)
They were kind of tightly coupled and we always needed PluginClient to
instantiate RemotePluginInstance, or PluginService to instantiate
LocalPluginInstance. But they do not have to.

It is kind of annoying if we need only instance parts, especially when
we are implementing a plugin wrapper. We want to use the instance API
without host API. Otherwise, we will have to concisely follow what the
instance API does e.g. nicely populate ports without `confirmPorts()`.

After adjusting the headers structure, there is no friend decls in those
host and instance classes anymore.