Added event receiver for unified inputs to aap::PluginInstance.
This is the first revision that could successfully dispatch Web UI inputs
to the PluginInstance on aapinstrumentsample (to some extent).

The existing ui-compose PluginDetails is not much capable of dealing with
the UI inputs (as it does not constantly processing audio) but the
parameter changes on the Web UI is reflected on the processing results.

There are more work to do, but we are getting able to finally close the
issue #34.