reverted strongly typed "priority" from parameters, and added full enum support.
Parameter priority is now part of parameter properties.
There are optional function to return parameter property in `parameters.h`.

Enums cannot be part of properties, so they are now explicitly added as
part of the API. It is implemented everywhere as long as I am aware of.

(They need to be implemented in aap-lv2 and aap-juce importers as well.)