Ensure that binder-client-as-plugin can also invoke extensions.
Formerly, it was regarded as it does not need to invoke get_extension()
because they could be invoked via AAPXS (it is how get_extension() is
implemented in this change after all). But when we commonize some code
in aap::PluginClient and aap::PluginService, namely when we populate
parameters, we are rather using AAP public API to achieve that.

This might not make *better* sense after all, but to match the implementation
to how it currently works, ensure that get_extension() also works at
the Binder client too.

It is part of the fix for issue #160.