[BREAKING CHANGE] fix #160 - implement dynamic parameter list via binder.
This involves breaking change in parameters extension, in that it does
not return pointer to `aap_parameter_info_t` and `aap_parameter_enum_t`
anymore. It was kind of API design flaw that those extension functions were
returning non-serializable values across Binder (we could still implement
such functions using potentially stale buffer, but that's awkward).

The corresponding extension URI is updated ("/v2" to "/v2.1").
Any involved code is changed for this API breakage (only in this repo so far).

This fix involves whole new changes in the premise in parameters extension
that it never needed Binder IPC - the dynamic parameter list was never
available without interaction between host and client.
In this changeset it is correctly implemented.

aapbarebonepluginsample is updated to dynamically populate parameter list,
and aapinstrumentsample is updated to not do that yet provide parameters v2.1
extension (it was useful to detect other kind of runtime crashes).
aapbarebonepluginsample is also configured to provide native (Compose) UI.

They all work fine.