Remove use of AudioPluginInstance class from some of public API surface.
AudioPluginInstance of current API does not expose sufficient functionality.
So we should replace it with NativeRemotePluginInstance of anything that
makes use of that class, but currently AudioPluginInstance class is used
by various essential API parts. They should be replaced by the thinner
native wrapper. So far remove some of those.