plugin-manager: port midi-device-service implementation to MidiSourceNode.
There are some additional changes on MIDI buffering etc. as well as some
additional mentions and formalization on MIDI events for extension control.
We need to replace "setCurrentPresetIndex()" call for Program Change.
Current implementation turned out to not care about that RT safety much.

Since it was not very easy to implement it, we will visit that later.