implement AAPXS SysEx8 generator support, up to the new plugin manager (WIP)
context: https://github.com/atsushieno/aap-core/issues/169

Unlike service side, the client side needs to handle it at app level.

It is still not verified, as it turned out that we do not have any feature
that triggers preset setter in the current manager UI. We need to add it.