for use in aap-juce, avoid extra reference to `<cmidi2.h>`.
JUCE module architecture is immature when it comes to header files resolution.
We will have to set additional include paths if the reference is <cmidi2.h>,
not "cmidi2.h". With CMake support, it could be just an external library.
Unfortunately we are not always CMake based - there are still some legacy
Producer based projects and they cannot resolve <cmidi2.h>.

But, in the first place, we already include "cmidi2.h" (note the quote vs.
bracket difference) and we do not need to include <cmidi2.h> at all.
So, instead of tweaking CMakeLists.txt and `*.jucer` files, we could add
hacky workarounds by copying #ifndef safeguard for duplicate includes.

(It is really a hacky workaround because cmidi2 is not really committed to
use the same definition symbol, and at any state it colud be changed.)