add asynchronous reply slot function to AAPXSClientInstance.
Now public API wise we have asynchronous AAPXS handler that invokes the newly
added `RemotePluginInstance::processExtensionReply()` which is stub for now.

We still need to design how thread dispatching works.
At this state, extension functions like `getPresetCount()` is defined as a
synchronous method that needs to block until it gets the result.
And that is added at `StandardExtensions` as synchronous functions, so they
need to be asynchronized.
That would lead to async enhancements at `PluginClientExtensionImplBase` class.