async-ify get_preset() and implement the rest of realtime AAPXS stuff.
Regarding these issues:

https://github.com/atsushieno/aap-core/issues/182

`get_preset()` now takes async callback arguments, so every relevant parts
needed to be updated too. This was a big chunk of breaking changes.

https://github.com/atsushieno/aap-core/issues/176

Since our AAPXS client handler API is not async-ified, it still works in
synchronous mode, but the extension functionality is basically back (in that
preset list retrieval is fully functional again).

We likely need to revamp the overall AAPXS stuff (especially the whole
foundation behind StandardExtensions is synchronous), but we recover the
basic functionality first in this commit.

https://github.com/atsushieno/aap-core/issues/169

The client side code is, as explained above, hacky overall, but there is
certain progress.