(kinda) bring back extension proxy object support in AAPXSDefinition.
I thought we could just eliminate strongly typed extension proxies from
AAPXSDefinition.
It turned out that binder-client-as-plugin still needs to implement
`get_extension()` function, and it needs an extension instance for each URI.

The problem was hidden because `get_extension()` in binder-client-as-plugin
was buggy and was trying to get the extension from AndroidAudioPluginHost,
whose extension retrieval was also buggy and returned plugin extensions.

Fortunately?, TypedAAPXS classes could be used to represent the extension
functions and they could be kinda populated with additional contexts (such
as AAPXSInitiatorInstance and AAPXSSerialization).