Remove requestId argument from the extension completion callbacks.
If a request ID should be passed, then the caller should pass it around at
the invocation too. But (so far) get_preset() does not, and it still passes
callbackContext. The request ID should be included within the context (or
it should become the entire context).

This fixes the "missing request ID" issue in aap-juce wrapper implementation.