Experimental changes for issue #194 repro:
Build and install to mavenLocal with: `./gradlew publishToMavenLocal`

(no `build` this time, as it will fail at aapinstrumentsample)

Build and run `aapinstumentsample` on Android Studio (any version; I used
Jellyfish)

Select the only one item in the list and tap "Start" -> "-- Presets --"

with ndk r23 (as indicated in gradle/libs.versions.toml), it shows the items.

with ndk r24 the UI freezes.