re: #194: we do not specialize particular source anymore - cmidi2.h has #pragma.
It seems we can specify -fvectorize without conditionally applying to the
sources, now that we have general vectorization avoidance in cmidi2.h (which
seemed to be the occurrence point of this issue).

I thought we could simply go back to -O2, but such a build results in the
same UI freeze problem, so I leave it as -O1 + a bunch of optimization options.