Binder: callback instance should be BnXXX class.
This fixes https://github.com/atsushieno/aap-core/issues/188 .

It was due to misuse of proxy classes - the service callback object should be
BnAudioPluginInterfaceCallback which works like a "Stub" class in Java.
We do not have to create and pass BpAudioPluginInterfaceCallback by ourselves
as some equivalent code should be automatically created as part of proxying.